Refactored MoneroPayCallbackServer
