+++ XMRpos/app/src/main/java/org/monerokon/xmrpos/data/remote/moneroPayCallback/MoneroPayCallbackManager.kt
@@ -1,0 +1,18 @@
+package org.monerokon.xmrpos.data.remote.moneroPayCallback
+
+import org.monerokon.xmrpos.data.remote.moneroPayCallback.model.PaymentCallback
+
+class MoneroPayCallbackManager(private val port: Int) {
+    private var callbackServer: MoneroPayCallbackServer? = null
+
+    fun startListening(onPaymentReceived: (PaymentCallback) -> Unit) {
+        if (callbackServer == null) {
+            callbackServer = MoneroPayCallbackServer(port, onPaymentReceived)
+        }
+    }
+
+    fun stopListening() {
+        MoneroPayCallbackServer.stopServer(callbackServer)
+        callbackServer = null
+    }
+}
