+++ XMRpos/app/src/main/java/org/monerokon/xmrpos/data/remote/moneroPayCallback/model/PaymentCallback.kt
@@ -1,0 +1,12 @@
+package org.monerokon.xmrpos.data.remote.moneroPayCallback.model
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class PaymentCallback(
+    val amount: PaymentCallbackAmount,
+    val complete: Boolean,
+    val description: String,
+    val created_at: String,
+    val transaction: PaymentCallbackTransaction
+)
