+++ XMRpos/app/src/main/java/org/monerokon/xmrpos/data/remote/moneroPayCallback/model/PaymentCallbackAmount.kt
@@ -1,0 +1,9 @@
+package org.monerokon.xmrpos.data.remote.moneroPayCallback.model
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class PaymentCallbackAmount(
+    val expected: Long,
+    val covered: PaymentCallbackCovered
+)
