+++ XMRpos/app/src/main/java/org/monerokon/xmrpos/data/remote/moneroPayCallback/model/PaymentCallbackCovered.kt
@@ -1,0 +1,9 @@
+package org.monerokon.xmrpos.data.remote.moneroPayCallback.model
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class PaymentCallbackCovered(
+    val total: Long,
+    val unlocked: Long
+)
